problem: not clear which rockets are testnet at `/buymerits`
resolve https://github.com/nostrocket/hypergolic/issues/82
